Theoretical Maximum Habitat

A list for each species containing an ordered vector of the maximum possible suitable area for each watershed

max_spawn_area

max_rear_area

Details

Suitable habitat modeling captures degraded existing habitat conditions. For the purposes of scenario development, we assumed that existing maximum habitat (which occurs at the most suitable flow conditions in the watershed) can be increased by a maximum of 100 feasibility of habitat restoration over twenty years as well as physical limitations (e.g. homes, levees, bridges, etc.) on habitat conditions.

Examples

# Subset by run to find max spawn and rear habitat area
R2Rscenario::max_spawn_area$FALL
#>     Upper Sacramento River             Antelope Creek 
#>                 541397.931                 123291.779 
#>               Battle Creek                 Bear Creek 
#>                   6815.051                  89034.974 
#>            Big Chico Creek                Butte Creek 
#>                  78367.408                  15007.990 
#>                Clear Creek           Cottonwood Creek 
#>                 101992.136                   7265.043 
#>                  Cow Creek                 Deer Creek 
#>                 408442.999                 193354.780 
#>                Elder Creek                 Mill Creek 
#>                  36958.952                 106149.647 
#>               Paynes Creek                Stony Creek 
#>                  61473.297                 135897.111 
#>               Thomes Creek Upper-mid Sacramento River 
#>                 195021.657                      0.000 
#>              Sutter Bypass                 Bear River 
#>                      0.000                 121201.151 
#>              Feather River                 Yuba River 
#>                 471825.938                 251529.370 
#> Lower-mid Sacramento River                Yolo Bypass 
#>                      0.000                      0.000 
#>             American River     Lower Sacramento River 
#>                2040494.872                      0.000 
#>            Calaveras River             Cosumnes River 
#>                  99942.916                 105369.628 
#>            Mokelumne River               Merced River 
#>                 112701.298                 343810.544 
#>           Stanislaus River             Tuolumne River 
#>                 256861.185                 384228.652 
#>          San Joaquin River 
#>                      0.000 
R2Rscenario::max_rear_area$FALL
#>     Upper Sacramento River             Antelope Creek 
#>                  597320.75                  407094.78 
#>               Battle Creek                 Bear Creek 
#>                   79205.70                  222688.41 
#>            Big Chico Creek                Butte Creek 
#>                  193828.01                  353798.99 
#>                Clear Creek           Cottonwood Creek 
#>                   88289.18                  495279.74 
#>                  Cow Creek                 Deer Creek 
#>                 1137373.05                  659839.85 
#>                Elder Creek                 Mill Creek 
#>                   91422.49                  263123.45 
#>               Paynes Creek                Stony Creek 
#>                  152043.39                  336306.84 
#>               Thomes Creek Upper-mid Sacramento River 
#>                  482477.84                 1109100.15 
#>              Sutter Bypass                 Bear River 
#>                       0.00                  384798.62 
#>              Feather River                 Yuba River 
#>                 4513105.74                  138520.26 
#> Lower-mid Sacramento River                Yolo Bypass 
#>                  750518.49                       0.00 
#>             American River     Lower Sacramento River 
#>                 2484444.21                  145451.48 
#>            Calaveras River             Cosumnes River 
#>                  132958.38                  749048.98 
#>            Mokelumne River               Merced River 
#>                 1656337.76                 3060926.48 
#>           Stanislaus River             Tuolumne River 
#>                  142007.84                 2815094.91 
#>          San Joaquin River 
#>                 1376900.51